Fixes json_test.go
The test to verify JSON encoding of a recursive object currently fails.  This modifies the Lua block to check the error message, similar to other json#encode failure assertions.

In the longer term this repository should be rebased off of github.com/layeh/gopher-json which seems to pass even with the previous assertion structure of `assert(json.encode(obj) == nil)`.